+++ src/lib/components/LexiconListViewHeader.svelte
